change code type from long to int, resolve problems with recyclerview and sort list by name

+++ app/src/main/java/com/example/pinapp/model/dao/Pin.kt
@@ -9,7 +9,6 @@ open class Pin: RealmObject {
      @PrimaryKey
      var id: ObjectId = ObjectId.invoke()
      var name: String = ""
-     var code: Long = -1
-
+     var code: Int = -1
 }
