+++ app/src/main/java/com/example/pinapp/model/domain/PinDomain.kt
@@ -5,5 +5,5 @@ import org.mongodb.kbson.ObjectId
 data class PinDomain(
     var id: ObjectId? = null,
     var name: String,
-    var code: Long
+    var code: Int
 )
